requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
requests-cache==1.1.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3